refactored folder structures, removed extra components
