+++ src/assets/styles/index.scss
@@ -52,7 +52,7 @@ body {
 	vertical-align: text-top;
 }
 
-@import '../../directives/tooltip/css/index.scss';
+@import '../../plugins/directives/tooltip/css/index.scss';
 
 .custom-tooltip {
 	background: #000;
