+++ src/boot/globalComponents.ts
@@ -1,0 +1,21 @@
+import BaseLayout from 'components/layouts/BaseLayout.vue';
+import Modal from 'components/Modal.vue';
+import Button from 'components/UIElements/Button.vue';
+import NewButton from 'components/UIElements/NewButton.vue';
+import DropdownMenu from 'components/UIElements/DropdownMenu.vue';
+import Loader from 'components/loaders/Loader.vue';
+import DayNightSwitch from 'components/UIElements/DayNightSwitch.vue';
+import AccountDropdown from 'components/UIElements/AccountDropdown.vue';
+import Alert from 'components/UIElements/Alert.vue';
+
+export default [
+	Alert,
+	BaseLayout,
+	DayNightSwitch,
+	Modal,
+	DropdownMenu,
+	Button,
+	NewButton,
+	Loader,
+	AccountDropdown,
+];
