+++ src/components/Confetti.vue
