+++ src/components/Slideout.vue
