+++ src/components/general/CategoryBadge.vue
