+++ src/components/general/Confirm.vue
