+++ src/pages/WorkspaceInvitation.vue
@@ -30,7 +30,7 @@
 </template>
 
 <script>
-	import Button from 'src/components/UIElements/Button.vue';
+	import Button from 'src/components/general/Button.vue';
 	import { getUser, updateUser } from 'src/actions/tmgr/user';
 	import { acceptWorkspaceInvitation } from 'src/actions/tmgr/workspaces';
