+++ src/plugins/directives/selectable/index.js
