+++ src/plugins/directives/selectable/selectable.js
