+++ src/plugins/directives/tooltip/directives/tooltip-directive.js
