+++ src/plugins/directives/tooltip/directives/tooltip.js
