--- src/utils/breadcrumbs/getProjectCategoriesBreadcrumbs.js
@@ -1,14 +1,0 @@
-export default function getProjectCategoriesBreadcrumbs(parentCategories) {
-	if (!parentCategories) {
-		return [];
-	}
-	const result = [];
-	parentCategories.map((item) =>
-		result.push({
-			label: item.title,
-			to: `/projects-categories/${item ? item.id + '/children' : ''}`,
-			payload: item,
-		}),
-	);
-	return result;
-}
